Se ha corregido el Bug que impedía usar el generador de assets aleatorios si se usaba /ban respondiendo un mensaje. Además, si se responde a un mensaje y el usuario ya estaba baneado, el bot responderá con un mensaje de error indicando que ya esta baneado

+++ src/commands/admin_commands/ban.rs
@@ -1,5 +1,4 @@
 use crate::commands::*;
-use crate::utils::ban_animation_generator;
 
 pub async fn ban_user(bot: Bot, msg: Message) -> ResponseResult<()> {
     // The function takes a bot and a message object, and returns a Result.
@@ -36,15 +35,56 @@ pub async fn ban_user(bot: Bot, msg: Message) -> ResponseResult<()> {
                 let is_admin_or_owner = chat_member.status().is_administrator() || chat_member.status().is_owner();
 
                 if is_admin_or_owner {
-                    bot.ban_chat_member(msg.chat.id, user.id).await?;
-                    let ban_msg = bot.send_message(msg.chat.id, format!("✅ @{} \\[`{}`\\] baneado", username_user, user.id)).reply_to_message_id(msg.id).await?;
 
-                    sleep(Duration::from_secs(5)).await;
-                    bot.delete_message(msg.chat.id, ban_msg.id).await?;
-                    bot.delete_message(msg.chat.id, msg.id).await?;
+                    let chat_member = bot.get_chat_member(msg.chat.id, user.id).await?;
+                    if chat_member.status().is_banned() {
+                        let err = bot.send_message(msg.chat.id, format!
+                        ("❌ {} Ya está baneado. Usa este comando solo para banear a alguien que no este baneado ", username_user))
+                            .reply_to_message_id(msg.id)
+                            .parse_mode(ParseMode::Html)
+                            .await?;
 
-                    // Choose a random ban animation to send.
-                    ban_animation_generator(bot, msg).await?;
+                        sleep(Duration::from_secs(5)).await;
+                        bot.delete_message(msg.chat.id, err.id).await?;
+                        bot.delete_message(msg.chat.id, msg.id).await?;
+
+                        return Ok(());
+                    } else {
+                        bot.ban_chat_member(msg.chat.id, user.id).await?;
+                        let mut rng: StdRng = SeedableRng::from_entropy();
+                        let random_number = rng.gen_range(0..=14);
+                        let file_names = [
+                            "1.gif", "2.gif", "3.gif", "4.gif", "5.gif",
+                            "6.gif", "7.gif", "8.gif", "9.gif", "10.gif",
+                            "11.gif", "12.mp4", "13.mp4", "14.mp4",
+                        ];
+                        let get_file_name = |index: usize| -> &'static str {
+                            file_names.get(index).unwrap_or_else(|| file_names.last().unwrap())
+                        };
+                        let file_path = format!("./assets/ban/{}", get_file_name(random_number));
+
+                        match file_path.ends_with(".gif") {
+
+                            true => {
+                                bot.send_animation(msg.chat.id, InputFile::file(file_path))
+                                    .caption(format!("✅ @{} [<code>{}</code>] baneado", username_user, user.id))
+                                    .parse_mode(ParseMode::Html)
+                                    .reply_to_message_id(msg.id)
+                                    .await?;
+                            }
+
+                            false => {
+                                bot.send_video(msg.chat.id, InputFile::file(file_path))
+                                    .caption(format!("✅ @{} <code>[{}</code>] baneado", username_user, user.id))
+                                    .parse_mode(ParseMode::Html)
+                                    .reply_to_message_id(msg.id)
+                                    .await?;
+                            }
+
+                        }
+
+                    }
+
                 } else {
                     // If the user is not an admin or owner, send an error message and delete this message in 5 seconds.
                     let err = bot.send_message(msg.chat.id, "❌ No tienes permisos para usar este comando").reply_to_message_id(msg.id).await?;
